requirements.txt
# Core Flask dependencies
Flask==2.3.3
Werkzeug==2.3.7

# Database and Storage
supabase==1.0.3
psycopg2-binary==2.9.9

# Environment and Configuration
python-dotenv==1.0.0

# Security
cryptography>=41.0.0

# File handling and utilities
Pillow>=10.0.0
python-magic>=0.4.27

# Development and debugging (optional)
flask-cors>=4.0.0

# Production server (optional)
gunicorn>=21.2.0

# Logging and monitoring (optional)
python-json-logger>=2.0.7
